requirements.txt
cython>=0.28
astropy>=3.0.4
backports-abc>=0.5
bokeh>=0.13
certifi>=2017.1.23
Django>=2.1
djangorestframework>=3.8.2
ephem>=3.7.6.0
futures>=3.0.5
h5py>=2.8
Jinja2>=2.10
MarkupSafe>=0.23
numpy>=1.15.1
pyparsing>=2.0.3
python-dateutil>=2.7.3
pytz>=2016.2
PyYAML>=3.11
requests>=2.10.0
scipy>=1.1
singledispatch>=3.4.0.3
six>=1.10.0
tornado>=4.3
pandas>=0.23.4

